config: Move `Server` struct to dedicated `config::server` module
